Woooh ... Before & After Analysis for different time series for overall sales look at different intervals. Using Generate_Series, inclusion for sum/case when for subquery inclusion on cte, and finally unique final row return for various case/when statements as well as percentage formatting and dealing with character type casting for operation capability and all case branch logic being of same type

+++ Sql_Challenges/5_DataMart/Data_Mart_CS.sql
@@ -216,7 +216,7 @@ FROM age_demographic_retail_sales
 SELECT 
   *,
   ROUND(
-    100 * (age_demo_sales / SUM(age_demo_sales) OVER())
+    100 * (age_demo_sales::NUMERIC / SUM(age_demo_sales) OVER())
   , 1) AS age_demo_percentage_ofsale
 FROM age_band_sales_rankings 
 ORDER BY demo_ageband_retail_sales_rankings;
